Change outside and inside logic of the component
